requirements.txt
Django
South
pytz
geopy
python3-memcached
psycopg2
decorator
python-social-auth
